fix: fix cropped output when using Animate CC's camera
Animate CC's camera translates the mask of its parent (if there
is one) so that the mask's (0, 0) point coincides with its center,
and we do set a mask on it, so we end up with cropped output.

By setting the mask's position to (0, 0) just before a frame is
rendered, this problem is avoided.

Closes #1.